boletofacilsdk/src/main/java/com/boletobancario/boletofacilsdk/model/entities/Payer.java
package com.boletobancario.boletofacilsdk.model.entities;

public class Payer extends Person {
	private String email;
	private String secondaryEmail;
	private String phone;

	public String getEmail() {
		return email;
	}

	public void setEmail(String email) {
		this.email = email;
	}

	public String getSecondaryEmail() {
		return secondaryEmail;
	}

	public void setSecondaryEmail(String secondaryEmail) {
		this.secondaryEmail = secondaryEmail;
	}

	public String getPhone() {
		return phone;
	}

	public void setPhone(String phone) {
		this.phone = phone;
	}
}
